requirements.txt
PyQt5
requests
streamlit
Pillow
python-dotenv
google-genai
google-generativeai
psycopg2-binary
mysql-connector-python